requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile - -o requirements.txt
neo4j==5.23.1
python-dotenv==1.0.1
pytz==2024.1
    # via neo4j
tqdm==4.66.5
